requirements.txt
gunicorn==20.1.0
Flask==2.2.2
mysql-connector==2.2.9